Updated the Interpolator with Ashi's feedback
The new Interpolator was changed to calculate with the analytical linear intrerpolator instead of using a linear regression on the points to make it faster. A superfluousInterpolatedRowCleanup was added to delete interpolated rows that were added unnecessarily by the main interpolator function due to the calculations made with both the min and max log ratios. In the main interpolator function, Clint had calculated each interpolated value independently. This process was vectorized to interpolate values with a list of abscissa values that need to be inserted across the YYY data. The function now returns the list of insertion indices.

For Test1 and Test2 results to be comparable to the old results, the function call had to be changed to addcleanupSuperfluousInterpolatedRows=False. Test3 was made as a copy of Test2, but with cleanupSuperfluousInterpolatedRows=True. The results for Test3 are also shown in the results by hand file.

The expected results files had to change due to the new method of interpolated and the returning of the insertion_indices. The new method of interpolation created insignificant numerical differences (to the order of 10^ -6) that flag the unit tester but are essentially   the same as the old one.