removeColumnsAtZeroOrBelowThreshold added
Added a function that will delete python array columns where all of the values are equal to zero or below a certain threshold. I wanted to include these in one function as it implies that zeros will be included as below the threshold if the threshold is grater than of equal to zero.