Add removeColumnsAtZeroOrBelowThreshold
Remove columns at zero or below threshold will delete python array columns (appearing as rows on spreadsheets) where all the values are either zero or below a threshold. The default of the function will remove zeros, but the optional arguments can be changed to delete columns below a threshold. Additionally, a startingRowIndex (appears as columns on spreadsheets) can be used to evaluate a subset of the original array. This becomes useful for evaluating a XYYY data set, where the X-rows should not be evaluated.