Renamed removeColumnsWithAllvaluesBelowZeroOrThreshold
The original purpose of this merge was to remove a print statement that I had left for debugging purposes, but I also made some changes according to previous comments from Ashi. In the XYYYDataFunctions module, I renamed the removeColumnsAtZeroOrBelowThreshold to removeColumnsWithAllvaluesBelowZeroOrThreshold to clarify the function. I also made the default threshold zero and removed the special case for zeros.