Updated the testing module for system agnoticism
Previously a lot of the tests were hardcoded with
Windows style paths. This commit changes those to
os package calls that should work on both Windows
and Unix systems (Mac and Linux).

There were a few changes required outside the testing
directory as well. In the user input file, default
user input file, and MSRESOLVE. It was
always similar issues, converting Windows style
paths to os package calls.

At this point all of the tests pass with the pytest driver
except for tests 4,5, and 6 in the ExractReferencePatternFromData
folder.

Prior to this commit they didn't run at all on Linux. When I ran
the tests on my Windows machine prior to the commit the same
three tests failed and the iterative analysis failed as well.

I am now going to test this commit on my windows machine.