Ensures all tests pass on Linux and Windows
Previously the three later tests in ExtractReferencePatterhFromData
did not pass on my Ubuntu installation. Additionally on my
Windows machine those same three tests and the IterativeAnalysis
test also failed.

The three tests in Extract... passed, for both Ubuntu and
Windows once I added a
relativeTolerance to the check_results call. I initially set it
as 1e-5. I have taken it to be as small as 1e-30 and the tests
still pass. Apparently any value at all (other than None) will work.
I did not investigate where this threshold was being used.

The IterativeAnalysis test actually wasn't failing on my Windows
machine but was throwing an exception before completion. It occured
in the MSRESOLVE main() function when

`importlib.import_module(...)`

is called. You can see that I changed the arguments slightly and
it now works fine. I find it pretty odd that the old arrangement
worked on Ubuntu and everyone else's Windows installations but
not on mine.

From what I can tell of the documentation on importlib the call
does exaclty the same thing as it was doing before. I checked to
be sure that it is actually loading the right module (i.e. the
proper global variables) and it seems to be.