Fixing Iterative w/ UserInput & Ionization Factors
Apparently the way the new user input structure is and the loading of ionization factors was no longer compatible with iterative (even though iterative was passing the unit test). I needed to move the loading of the ionization factors above iterative, to before the directory change occurred. And, needed to change the if statement to a try statement, as one way to get the desired behavior.