Improving autoscaling to 'ignore' non-numbers
Previously, if there were blanks in the experimental data csv file, this caused autoscaling to crash because the blanks were different from 0 but still resulted in division by 0. Now blanks are 'ignored' when deciding the scaling factor by autoscaling.
I also improved one warning message to be more clear (regarding use of 'referenceValueThreshold' feature if SLS is not able to solve).