Improvement to generalize concentrations finder
Concentration finder was already working with multiple molecules under certain scenarios.  However, there was a logic flaw such that under certain circumstances the mass indices would be wrong when the reference file and the experimental collection file each had extra masses. This should never have been a problem, and was because a mass index was being pulled from the experimental array when it should have been pulled from the reference data array the logic flaw is fixed.