making ionization related variable names consistent & fixing object casting and logic as needed
This change was particularly challenging. Many changes had to be done at one time to prevent breaking of the desired behavior. It was not possible to do this solely incrementally, and not easy to find a suitable order for doing these changes. It is probably better to do all at once now that it is known what they are.  In addition to the below variable naming, these variables were also added in a few new lines.

knownMoleculesIonizationTypes -- > moleculeIonizationType
SourceOfIonizationData --> sourceOfIonizationData
sourceOfIonizationDataList --> sourceOfIonizationData
ionizationEfficienciesList --> relativeIonizationEfficiencies
ionizationEfficienciesSourcesList --> sourceOfIonizationData