Allowed full run to occur
The 1butanal is basically nothing, as it should be.
If dividing the 3 concentrations of the 2 carbon species by the smallest concentration one which is ethyne, the relative concentrations become:
ethyne:1, ethene:2.97, ethane:6.72
Note that this is different from the coefficients we started with of ethane:ethene:ethyne of 10:3:1. . This is fine:  what we started with was multiplying the patterns, which is different from the concentrations themselves.  In fact, the 6.72 number must have been the original answer anyway, barring any effect from the butanal, because the butanal is the only changed molecule in the pattern.