Uncertainties Edits. seems to be working with TuningCorrectorGasMixture
I made a bunch of code additions to TuningCorrectorGasMixture to add uncertainties calculation.  I also needed to change extendReferencePattern to return the reference slice that was used during the extension so that I could get the uncertainties associated with the slice.  The uncertainties calculations should probably be functionalized since it is occurring in several places now.

It seems to be working because tor the 3 molecules from the original pattern, their sclaed concentrations and the uncertainties in the scaled concentrations are nearly the same as before. The new molecules has a high relative uncertainty (e.g. 1000 percent) but it is still a lower absolute uncertainty relative to the concentrations of the other molecules.