Fixed the ConcentrationFinder bug
The problem was that in ratioFinder there was a numpy "Where" that required a numpy array and was receiving a list. I changed the where to force an array casting in that line, and did the same thing in another part of the code that also had a where statement.  Now, all unit tests pass except for tuning corrector test 6 which has not been finished yet.