I think I found the problem. I think it is reference pattern molecules order.
In the files ExportedReferencePatternOriginalForCorrectionValuesStandardTuning.csv and ExportedReferencePatternStandardForCorrectionValuesMixedStandardTuning.csv the molecules are in different orders. That means the standardized_reference_patterns and other arrays are in different orders.  It seems that I will need to make a function that re-orders the molecules in a reference pattern.  In order to make that easy, I will probably make a function that takes the desired order of molecules as a list, and then makes a reference pattern for each molecule.  Then it can add the patterns together one at a time until the right order is achieved.