Fixing Mixed Reference Patterns Bug, update userInputValidityFunctions.py
There was a logic error bug in UserInputValidity functions: When 'external' became the default for tuning correction in the last year, the createMixedReferencePattern functionalitiy changed and the userInputValidity functions was not updated to have the correct behavior for creating mixed reference patterns. userInputValidity functions was turning off creation of the mixed reference pattern.